app/requirements.txt
Django==2.0.3
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.3
django-multiselectfield==0.1.8
messytables==0.15.2
certifi==2018.4.16
chardet==3.0.4
Django==2.0.3
django-multiselectfield==0.1.8
html5lib==1.0.1
idna==2.6
json-table-schema==0.2.1
lxml==4.2.1
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-dateutil==2.7.2
python-dotenv==0.8.2
python-magic==0.4.15
pytz==2018.3
requests==2.18.4
six==1.11.0
urllib3==1.22
webencodings==0.5.1
xlrd==1.1.0
